enable auditing and fix datatype on entities

+++ src/main/java/com/amit/practice/repository/ProductImageRepository.java
@@ -3,5 +3,5 @@ package com.amit.practice.repository;
 import com.amit.practice.model.ProductImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
 }
