+++ src/main/java/com/amit/practice/repository/ProductReviewRepository.java
@@ -3,5 +3,5 @@ package com.amit.practice.repository;
 import com.amit.practice.model.ProductReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductReviewRepository extends JpaRepository<ProductReview, Integer> {
+public interface ProductReviewRepository extends JpaRepository<ProductReview, Long> {
 }
